fix!: Move backwards-compatibility hacks out of blockly.js
Move accessor properties for Blockly.alert, .confirm,
.mainWorkspace, .prompt, .selected, .HSV_SATURATION and
.HSV_VALUE, as well as the hack to allow loading of messages
via <script> tags, from core/blockly.js to core/main.js,
which becomes the entrypoint for the first chunk when
compiled.

BREAKING CHANGE: The aforementioned properties / hack are no
longer available in uncompiled mode (or when using advanced
compilation, unless also compiling against main.js.)
